Move `SetOnceValidationRenderer` under the `setonce` package
